requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
numpy==1.24.4
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
pyTelegramBotAPI==4.12.0
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.3
